WORKS but caching doesn't

+++ query_cache/cache_request.go
@@ -18,7 +18,7 @@ type CacheRequest struct {
 
 	resultKeyRoot string
 	pageCount     int64
-	rowCount      int64
+	rowCount      int
 	StreamContext context.Context
 }
